feat : 옵션 최소 수량 조건을 위한 로직 추가
옵션의 수량이 하나일 때는 삭제되지 않도록 하였다

+++ src/main/java/gift/service/OptionService.java
@@ -87,6 +87,8 @@ public class OptionService {
      */
     @Transactional
     public void delete(Long id){
-        optionRepository.deleteById(id);
+        Option savedOption = optionRepository.findById(id).orElseThrow(NoSuchFieldError::new);
+        if(savedOption.getQuantity() != 1)
+            optionRepository.deleteById(id);
     }
 }
